Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    int a,b,c;
-    double average;
+    float a,b,c;
+    float average;
     scanf("%d%d%d",&a,&b,&c);
     average=(a+b+c)/3;
     printf("Average: %.2lf",average);
